# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/leads": {
      "get": {
        "operationId": "LeadsController_findAll",
        "summary": "List all leads",
        "description": "Returns a paginated list of leads assigned to or visible to the current user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of leads retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ]
      },
      "post": {
        "operationId": "LeadsController_create",
        "summary": "Create new lead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lead created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        },
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}": {
      "get": {
        "operationId": "LeadsController_findOne",
        "summary": "Get lead by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Lead not found"
          }
        },
        "tags": [
          "Leads"
        ]
      },
      "patch": {
        "operationId": "LeadsController_updateLead",
        "summary": "Update lead details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lead updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}/status": {
      "patch": {
        "operationId": "LeadsController_updateStatus",
        "summary": "Update lead status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/bulk": {
      "post": {
        "operationId": "LeadsController_bulkUpdate",
        "summary": "Bulk update leads",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateLeadsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leads updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{id}/convert": {
      "post": {
        "operationId": "LeadsController_convert",
        "summary": "Convert lead to project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lead converted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Contact mapping required before conversion"
          }
        },
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/contacts": {
      "get": {
        "operationId": "ContactsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "ContactsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/documents/upload": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/documents/{id}/approve": {
      "post": {
        "operationId": "DocumentsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "CRM Backend API",
    "description": "The core API for CRM Lead Management and Project Coordination.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crm_session",
        "description": "Session cookie (HttpOnly)"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "rid": {
            "type": "string",
            "example": "s-success-rid"
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "type": "object",
            "example": {
              "timestamp": "2026-03-28T07:44:00Z"
            }
          }
        },
        "required": [
          "rid",
          "statusCode",
          "data",
          "message",
          "meta"
        ]
      },
...
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Interior design for 3BHK",
            "description": "The title of the lead"
          },
          "description": {
            "type": "string",
            "example": "Client looking for premium villa fit-out"
          },
          "source": {
            "type": "string",
            "example": "website",
            "description": "Source of the lead"
          },
          "status": {
            "type": "string",
            "example": "new",
            "description": "Current status of the lead"
          },
          "estimatedValue": {
            "type": "number",
            "example": 500000,
            "description": "Projected value in INR"
          },
          "contactId": {
            "type": "string",
            "example": "uuid",
            "description": "Optional linked contact ID"
          },
          "assignedTo": {
            "type": "string",
            "example": "uuid",
            "description": "Optional assigned user ID"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateLeadDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Interior design for 3BHK",
            "description": "The title of the lead"
          },
          "description": {
            "type": "string",
            "example": "Client looking for premium villa fit-out"
          },
          "source": {
            "type": "string",
            "example": "website",
            "description": "Source of the lead"
          },
          "status": {
            "type": "string",
            "example": "new",
            "description": "Current status of the lead"
          },
          "estimatedValue": {
            "type": "number",
            "example": 500000,
            "description": "Projected value in INR"
          },
          "contactId": {
            "type": "string",
            "example": "uuid",
            "description": "Optional linked contact ID"
          },
          "assignedTo": {
            "type": "string",
            "example": "uuid",
            "description": "Optional assigned user ID"
          }
        }
      },
      "UpdateLeadStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "negotiating",
            "description": "New status for the lead"
          },
          "reason": {
            "type": "string",
            "example": "Client requested a discount"
          },
          "followUpAt": {
            "type": "string",
            "example": "2026-04-01",
            "description": "Next follow-up date"
          }
        },
        "required": [
          "status"
        ]
      },
      "BulkUpdateLeadsDto": {
        "type": "object",
        "properties": {
          "ids": {
            "example": [
              "uuid1",
              "uuid2"
            ],
            "description": "List of Lead IDs to update",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "example": "qualified"
          },
          "assignedTo": {
            "type": "string",
            "example": "uuid-user"
          }
        },
        "required": [
          "ids"
        ]
      },
      "ConvertLeadDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "example": "uuid-template",
            "description": "Optional project template ID to clone phases/tasks"
          }
        }
      },
      "CreateContactDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "example": "uuid-account"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "altPhone": {
            "type": "string",
            "example": "+1234567891"
          },
          "designation": {
            "type": "string",
            "example": "Senior Architect"
          },
          "companyName": {
            "type": "string",
            "example": "Design co."
          },
          "category": {
            "type": "string",
            "example": "architect",
            "enum": [
              "individual",
              "architect",
              "pmc",
              "vendor"
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "example": true
          },
          "address": {
            "type": "string",
            "example": "12 Crescent Road"
          },
          "city": {
            "type": "string",
            "example": "Gurugram"
          },
          "state": {
            "type": "string",
            "example": "Haryana"
          },
          "country": {
            "type": "string",
            "example": "India"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "notes": {
            "type": "string",
            "example": "Key decision maker"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "UpdateContactDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "example": "uuid-account"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "altPhone": {
            "type": "string",
            "example": "+1234567891"
          },
          "designation": {
            "type": "string",
            "example": "Senior Architect"
          },
          "companyName": {
            "type": "string",
            "example": "Design co."
          },
          "category": {
            "type": "string",
            "example": "pmc",
            "enum": [
              "individual",
              "architect",
              "pmc",
              "vendor"
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "example": true
          },
          "address": {
            "type": "string",
            "example": "12 Crescent Road"
          },
          "city": {
            "type": "string",
            "example": "Gurugram"
          },
          "state": {
            "type": "string",
            "example": "Haryana"
          },
          "country": {
            "type": "string",
            "example": "India"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "notes": {
            "type": "string",
            "example": "Key decision maker"
          }
        }
      },
...
      "UploadDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "project-brief.pdf"
          },
          "entityType": {
            "type": "string",
            "example": "projects",
            "description": "Preferred runtime entity or legacy module name"
          },
          "moduleName": {
            "type": "string",
            "example": "projects",
            "description": "Legacy alias kept for compatibility"
          },
          "entityId": {
            "type": "string",
            "example": "uuid"
          },
          "category": {
            "type": "string",
            "example": "design"
          },
          "versionLabel": {
            "type": "string",
            "example": "v1.0"
          },
          "approvalRequired": {
            "type": "boolean",
            "example": true
          },
          "shareMode": {
            "type": "string",
            "example": "internal",
            "enum": [
              "internal",
              "external",
              "restricted"
            ]
          },
          "description": {
            "type": "string",
            "example": "Initial release package"
          },
          "tags": {
            "example": [
              "design",
              "client"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApproveDocumentDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "example": "approved",
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "comment": {
            "type": "string",
            "example": "Approved after final review"
          },
          "versionLabel": {
            "type": "string",
            "example": "v1.1"
          },
          "parentDocumentId": {
            "type": "string",
            "example": "uuid-parent-document"
          }
        }
      },
...
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "isRead": {
            "type": "boolean",
            "example": true
          },
          "readAt": {
            "type": "string",
            "example": "2026-03-28T10:00:00.000Z"
          },
          "note": {
            "type": "string",
            "example": "dismissed from inbox"
          }
        }
      },
...
    }
  }
}